add logic to boss health and fix truncation bug

+++ Assets/Scripts/Player/PlayerData.cs
@@ -9,7 +9,7 @@ public class PlayerData : ScriptableObject
 {
     public int coin_counter;
 
-    public int total_coins;
+    public float total_coins;
 
     public int death_counter;
